Builds installation packages for COM objects.
Unfortunately, I am unable to automate testing of the COM installation.  What I do not understand is I am unable to script this COM msi. If I attempt to script the COM msi, the COM dll is never installed. I am speculating, that the custom action for the TARGETDIR is not functioning. If I use the mouse to double click on the msi, the privileges are raised to administrative privileges and the dll installs fine and the COM object is correctly installed.